+++ jobbergate-api/jobbergate_api/storage.py
@@ -13,8 +13,8 @@ import sqlalchemy
 from asyncpg.exceptions import UniqueViolationError
 from fastapi.exceptions import HTTPException
 from loguru import logger
-from sqlalchemy import Column, or_
-from sqlalchemy.sql.expression import BooleanClauseList, UnaryExpression
+from sqlalchemy import Column, Enum, or_
+from sqlalchemy.sql.expression import BooleanClauseList, Case, UnaryExpression
 from starlette import status
 from yarl import URL
 
@@ -65,11 +65,33 @@ def search_clause(
     return or_(*[field.ilike(f"%{term}%") for field in searchable_fields for term in search_terms.split()])
 
 
+def _build_enum_sort_clause(sort_column: Column, sort_ascending: bool) -> Case:
+    """
+    Build a Case statement that can be used as a sort clause for an enum column.
+
+    SQLAlchemy will not sort enums alphabetically by default, but rather by creation order.  Thus, we have to
+    force alphabetical sorting using a CASE clause. The logic here is strange. Basically, we have to provide a
+    lookup and then a string sort value. If sort order is ascending, each enum value's sort value is itself.
+    Otherwise, the sort value is it's opposite in the list. For example::
+
+        Given the enum (GAMMA, ALPHA, DELTA, BETA)
+        The default sort order used by SQLAlchemy is (GAMMA, ALPHA, DELTA, BETA)
+        Mapping used for ascending sort: {ALPHA: ALPHA, BETA: BETA, DELTA: DELTA, GAMMA: GAMMA}
+        Mapping used for descending sort: {ALPHA: GAMMA, BETA: DELTA, DELTA: BETA, GAMMA: ALPHA}
+
+    To understand this more fully, start with this SO question: https://stackoverflow.com/a/23618085/642511
+    """
+    assert isinstance(sort_column.type, Enum)
+    sorted_values = sorted(sort_column.type.enums)
+    sort_tuple = zip(sorted_values, sorted_values if sort_ascending else reversed(sorted_values))
+    return sqlalchemy.case(dict(sort_tuple), value=sort_column)
+
+
 def sort_clause(
     sort_field: str,
     sortable_fields: typing.List[Column],
     sort_ascending: bool,
-) -> typing.Union[Column, UnaryExpression]:
+) -> typing.Union[Column, UnaryExpression, Case]:
     """
     Create a sort clause given a sort field, the list of sortable fields, and a sort_ascending flag.
     """
@@ -81,8 +103,10 @@ def sort_clause(
             status_code=status.HTTP_400_BAD_REQUEST,
             detail=f"Invalid sorting column requested: {sort_field}. Must be one of {sort_field_names}",
         )
-    sort_column: typing.Union[Column, UnaryExpression] = sortable_fields[index]
-    if not sort_ascending:
+    sort_column: typing.Union[Column, UnaryExpression, Case] = sortable_fields[index]
+    if isinstance(sort_column, Column) and isinstance(sort_column.type, sqlalchemy.Enum):
+        sort_column = _build_enum_sort_clause(sort_column, sort_ascending)
+    elif not sort_ascending:
         sort_column = sort_column.desc()
     return sort_column